# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Internet Bank Backend API
  version: 1.1.0
  description: |
    Актуальная HTTP-спецификация backend-сервисов интернет-банка через API Gateway.
    Все запросы отправляются на один базовый URL.
servers:
  - url: http://localhost:8080
    description: API Gateway
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreditTariff:
      type: object
      required: [id, name, rate]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        rate:
          type: number
          format: double
        min_amount:
          type: number
          format: double
        max_amount:
          type: number
          format: double
    Credit:
      type: object
      required: [id, client_id, account_id, tariff_id, amount, remaining, rate, daily_payment, issued_at, status]
      properties:
        id:
          type: string
          format: uuid
        client_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        tariff_id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        remaining:
          type: number
          format: double
        rate:
          type: number
          format: double
        daily_payment:
          type: number
          format: double
        issued_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, paid, overdue]
    CreditOverdue:
      type: object
      required: [credit_id, client_id, expected_paid, actual_paid, overdue_amount, overdue_payments, minute_payment]
      properties:
        credit_id:
          type: string
          format: uuid
        client_id:
          type: string
          format: uuid
        expected_paid:
          type: number
          format: double
        actual_paid:
          type: number
          format: double
        overdue_amount:
          type: number
          format: double
        overdue_payments:
          type: integer
        minute_payment:
          type: number
          format: double
    CreditRating:
      type: object
      required: [client_id, score, risk_level, overdue_amount, overdue_count]
      properties:
        client_id:
          type: string
          format: uuid
        score:
          type: integer
        risk_level:
          type: string
        overdue_amount:
          type: number
          format: double
        overdue_count:
          type: integer
...
    TariffListResponse:
      type: object
      required: [tariffs, pageNumber, pageQuantity]
      properties:
        tariffs:
          type: array
          items:
            $ref: '#/components/schemas/CreditTariff'
        pageNumber:
          type: integer
        pageQuantity:
          type: integer
    CreditListResponse:
      type: object
      required: [credits, pageNumber, pageQuantity]
      properties:
        credits:
          type: array
          items:
            $ref: '#/components/schemas/Credit'
        pageNumber:
          type: integer
        pageQuantity:
          type: integer
...
    IssueCreditRequest:
      type: object
      required: [client_id, account_id, tariff_id, amount]
      properties:
        client_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        tariff_id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
    RepayCreditRequest:
      type: object
      required: [amount, account_id]
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
        account_id:
          type: string
          format: uuid
    CreateTariffRequest:
      type: object
      required: [name, rate, min_amount, max_amount]
      properties:
        name:
          type: string
        rate:
          type: number
          format: double
        min_amount:
          type: number
          format: double
        max_amount:
          type: number
          format: double
...
paths:
...
  /tariffs:
    post:
      tags: [Tariffs]
      summary: Создать тариф кредита (только сотрудник)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTariffRequest'
      responses:
        '201':
          description: Тариф создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditTariff'
    get:
      tags: [Tariffs]
      summary: Список тарифов
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 50
            minimum: 1
      responses:
        '200':
          description: Список тарифов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TariffListResponse'
  /credits:
    post:
      tags: [Credits]
      summary: Выдать кредит
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueCreditRequest'
      responses:
        '201':
          description: Кредит выдан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credit'
    get:
      tags: [Credits]
      summary: Список кредитов клиента
      parameters:
        - in: query
          name: client_id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 50
            minimum: 1
      responses:
        '200':
          description: Список кредитов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditListResponse'
  /credits/{creditId}:
    get:
      tags: [Credits]
      summary: Получить кредит по ID
      parameters:
        - in: path
          name: creditId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Кредит
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credit'
  /credits/{creditId}/repay:
    post:
      tags: [Credits]
      summary: Погасить кредит
      parameters:
        - in: path
          name: creditId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepayCreditRequest'
      responses:
        '200':
          description: Обновленный кредит
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credit'
  /credits/{creditId}/overdue:
    get:
      tags: [Credits]
      summary: Просрочка по конкретному кредиту
      parameters:
        - in: path
          name: creditId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Данные по просрочке
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditOverdue'
  /clients/{clientId}/credit-rating:
    get:
      tags: [Credits]
      summary: Кредитный рейтинг клиента
      parameters:
        - in: path
          name: clientId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Рейтинг клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditRating'
...